Fix error handling and parameter passing
Due to the new link placement there was no error handling, i.e. if a
script exited with a non zero exit code the whole process was not
terminated.

This is now fixed by adding a pipe with exit to the script call in the
link script, so that if any error occurs in the script this is passed on
to the caller.

Furthermore the arguments to the prepare-commit-msg script weren't
passed correctly. Now the whole argument array is passed to the script.

Also fixed formatting in the scripts.

Signed-off-by: Daniel Hiller <daniel.hiller.1972@gmail.com>